Revert "bugfix: valueops.toint". Maybe next time.
This reverts commit 3db5bcaf5ddd98f90135ea35c9bbb9b925557218.

+++ dev-clean/src/main/scala/sai/llsc/states/SymExeState.scala
@@ -99,7 +99,7 @@ trait SymExeDefs extends SAIOps with StagedNondet {
 
   type IntData = Long
   type BlockLabel = Int
-  type Addr = Long
+  type Addr = Int
   type PC = Set[SMTBool]
   type E = State[Rep[SS], *] ⊗ (Nondet ⊗ ∅)
 
@@ -197,7 +197,7 @@ trait SymExeDefs extends SAIOps with StagedNondet {
     def apply(i: Rep[Long])(implicit d: DummyImplicit): Rep[Value] = IntV(i, DEFAULT_INT_BW)
     def apply(i: Rep[Long], bw: Int)(implicit d: DummyImplicit): Rep[Value] =
       "make_IntV".reflectMutableWith[Value](i, bw)
-    def unapply(v: Rep[Value]): Option[(Long, Int)] = Unwrap(v) match {
+    def unapply(v: Rep[Value]): Option[(Int, Int)] = Unwrap(v) match {
       case Adapter.g.Def("make_IntV", Backend.Const(v: Int)::Backend.Const(bw: Int)::_) =>
         Some((v, bw))
       case Adapter.g.Def("make_IntV", Backend.Const(v: Long)::Backend.Const(bw: Int)::_) =>
@@ -241,13 +241,13 @@ trait SymExeDefs extends SAIOps with StagedNondet {
   implicit class ValueOps(v: Rep[Value]) {
     // TODO: optimization like int
     def loc: Rep[Addr] = "proj_LocV".reflectWith[Addr](v)
-    def int: Rep[Long] = v match {
+    def int: Rep[Int] = v match {
       case IntV(n, bw) => unit(n)
-      case _ => "proj_IntV".reflectWith[Long](v)
+      case _ => "proj_IntV".reflectWith[Int](v)
     }
     def float: Rep[Float] = "proj_FloatV".reflectWith[Float](v)
     def kind: Rep[Int] = "proj_LocV_kind".reflectWith[Int](v)
-    def structAt(i: Rep[Long]) = "structV_at".reflectWith[Value](v, i)
+    def structAt(i: Rep[Int]) = "structV_at".reflectWith[Value](v, i)
     def apply(s: Rep[SS], args: Rep[List[Value]]): Rep[List[(SS, Value)]] = {
       Unwrap(v) match {
         case Adapter.g.Def("llsc-external-wrapper", Backend.Const("noop")::Nil) =>
